Catched 404 not found when id is missing when try to delete it

+++ server/src/main/java/com/pcatalog/pcatalog/services/products/ProductsServiceImpl.java
@@ -30,6 +30,9 @@ public class ProductsServiceImpl implements ProductsService {
     @Override
     public Product deleteProduct(Long id) {
         Product product = productRepository.findByRecordId(id);
+        if (product == null){
+            return null;
+        }
         productRepository.delete(product);
         return product;
     }
